Moved IsTessOffChip, NggControl and GsOnChip into middle-end
Part of a series of patches to clean up the middle-end interface.

This commit moves NggControl and GsOnChip into PipelineState, and
the functions that set them up into PatchResourceCollect.
