+++ patch/gfx9/llpcNggPrimShader.h
@@ -35,12 +35,15 @@
 #include "llpc.h"
 #include "llpcInternal.h"
 #include "llpcNggLdsManager.h"
+#include "llpcPipelineState.h"
 
 namespace Llpc
 {
 
 class Context;
+struct NggControl;
 class NggLdsManager;
+class PipelineState;
 
 // Represents exported data used in "exp" instruction
 struct ExpData
@@ -56,7 +59,7 @@ struct ExpData
 class NggPrimShader
 {
 public:
-    NggPrimShader(Context* pContext);
+    NggPrimShader(PipelineState* pPipelineState);
     ~NggPrimShader();
 
     llvm::Function* Generate(llvm::Function* pEsEntryPoint,
@@ -210,8 +213,9 @@ private:
 
     static const uint32_t NullPrim = (1u << 31); // Null primitive data (invalid)
 
-    Context*        m_pContext; // LLPC context
-    GfxIpVersion    m_gfxIp;    // Graphics IP version info
+    PipelineState*  m_pPipelineState; // Pipeline state
+    Context*        m_pContext;       // LLPC context
+    GfxIpVersion    m_gfxIp;          // Graphics IP version info
 
     const NggControl* m_pNggControl;  // NGG control settings
